Add rm extra file and rename main file
