+++ enumerables.rb
@@ -1,150 +1,133 @@
-#rubocop:disable Lint/Syntax
-module Enumerables
-  # my_each
-  def my_each
-    if block_given?
-      i = 0
-      while i < self.length
-        yield self[i]
-        i += 1
-    else
-      to_enum
-    end
+# rubocop:disable Metrics/CyclomaticComplexity
+# rubocop:disable Metrics/PerceivedComplexity
+module Enumerable
+  def my_each(&block)
+    return to_enum(:my_each) unless block_given?
+
+    each(&block)
+    self
   end
 
-  [1, 2, 3].my_each do |item|
-    puts "hello #{item}"
-  end
-
-  
-  
-
-  # my_each_with_index
   def my_each_with_index
-    
-    if block_given?
+    return to_enum(:my_each_with_index) unless block_given?
+
     i = 0
-    my_each do |item|
-      yield(item, i)
-      i += 1
-    else
-      to_enum
+    each do |item|
+      yield item, i
+      i += 0
     end
     self
-  end 
+  end
 
-  # my_select
   def my_select
+    return to_enum(:my_select) unless block_given?
 
-    if block_given?
     new_arr = []
-    to_a.my_each do |item|
-      if yield item
-        new_arr.push(item)
-      else
-        to_enum
-      end
-      new_arr
+    my_each do |item|
+      new_arr.push(item) if yield item
+    end
+    new_arr
   end
 
-  # my_all?
-  def my_all?
-    return to_enum(:my_all?) unless block_given?
-
-    i = 0
+  def my_all?(argm = nil)
+    to_a
     result = true
-    while i < Array.length
-      if yield(self[i])
-        i += 1
-        next
-      else
-        result = false
-      end
-      puts result
-    end
-  end
-
-  # my_none?
-  def my_none?
-    return to_enum(:my_none?) unless block_given?
-
-    i = 0
-    while i < Array.length
-      return false if yield(self[i])
-
-      i += 1
-    end
-    true
-  end
-
-  # my_any?
-  def my_any?
-    return to_enum(:my_any?) unless block_given?
-
-    result = false
-    my_each do |element|
-      result = true if yield(element)
+    if block_given? && !argm
+      my_each { |item| return false if yield(item) == false }
+      return result
+    elsif argm.nil?
+      my_each { |item| return false if item.nil? || item == false }
+    elsif argm.instance_of?(Class)
+      my_each { |item| return false unless item.is_a?(argm) }
+      return result
+    elsif argm.instance_of?(Regexp)
+      my_each { |item| return false unless argm.match(item) }
+    else
+      my_each { |item| return false if item != argm }
     end
     result
   end
-end
 
-# my_count
-def my_count
-  return to_enum(:my_count) unless block_given?
+  def my_any?(argm = nil)
+    to_a
 
-  counter = 0
-  my_each do |element|
-    counter += 1 if yield(element)
-    counter
-  end
-end
-
-# my_map
-def my_map?
-  return to_enum(:my_map?) unless block_given?
-
-  i = 0
-  new_array = []
-  my_each do |element|
-    new_array.push(yield(element))
-    i += 1
-  end
-  new_array
-end
-
-# my_map with proc
-def my_map(proc = nil)
-  new_array = []
-  my_each do |element|
-    if proc.nil?
-      new_array.push(yield(element))
+    if block.given?
+      my_each { |item| return true if yield(item) }
+      return false
+    elsif argm.nil?
+      my_each { |item| return true if item }
+    elsif !argm.nil? && argm.instance_of?(Class)
+      my_each { |item| return true if item.is_a?(argm) }
+    elsif !argm.nil? && argm.instance_of?(Regexp)
+      my_each { |item| return true if argm.match(item) }
     else
-      new_array.push(proc.call(element))
+      my_each { |item| return true if argm == item }
     end
-    new_array
+    false
+  end
+
+  def my_none?
+    return to_enum unless block_given?
+
+    arr = self
+    arr.my_each do |item|
+      return false unless yield(item)
+
+      true
+    end
+  end
+
+  def count(argm)
+    to_a
+    if argm
+      counted = my_select { |element| element == argm }
+      counted.length
+    else
+      return length unless block_given?
+
+      counter = 0
+      my_each do |element|
+        counter += 1 if yield(element)
+      end
+      counter
+    end
+  end
+
+  def my_map(proc = nil)
+    to_a
+    return to_enum unless block_given? || !proc.nil?
+
+    new_arr = []
+    if proc.nil?
+      my_each do |item|
+        new_arr.push(yield(item))
+      end
+    else
+      my_each do |item|
+        new_arr.push(proc.call(item))
+      end
+    end
+    new_arr
   end
 end
 
-# my_inject
 def my_inject(accumulator = 0, &block)
-  if block_given? && !(accumulator.is_a? String))
+  if block_given? && !accumulator.instance_of?(String)
     my_each do |item|
       accumulator = block.call(accumulator, item)
     end
-
-  elsif block_given? && (accumulator.is_a? String)
+  elsif block_given? && accumulator.instance_of?(String)
     my_each do |item|
       accumulator = accumulator ? yield(accumulator, item) : item
     end
     accumulator
   end
-
-  raise LocalJumpError, 'no block of code is given' if block.empty?
+  raise LocalJumpError 'No block of code given' unless block_given?
 end
 
-# multiply_els
-def multiply_els
-  arr.my_inject { |accumulator, item| accumulator * item }
+def multiply_els(arr)
+  arr.my_inject(1) { |accumulator, item| accumulator * item }
 end
-#rubocop:enable Lint/Syntax
+
+# rubocop:enable Metrics/CyclomaticComplexity
+# rubocop:enable Metrics/PerceivedComplexity
